requirements.txt
# AI và Machine Learning
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Excel Operations
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Web Framework
streamlit>=1.28.0
streamlit-extras>=0.3.0

# Configuration
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Utilities
click>=8.1.0
rich>=13.7.0
pydantic>=2.5.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
